requirements.txt
eth-tester
py-solc-x
pytest
web3
